src/main/java/by/bestlunch/web/controller/user/ProfileRestController.java
package by.bestlunch.web.controller.user;

import by.bestlunch.persistence.model.User;
import by.bestlunch.security.AuthorizedUser;
import by.bestlunch.util.UserUtil;
import by.bestlunch.validation.view.ErrorSequence;
import by.bestlunch.web.dto.UserDto;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;
import java.net.URI;

@RestController
@RequestMapping(ProfileRestController.REST_URL)
public class ProfileRestController extends AbstractUserController {

    static final String REST_URL = "/rest/profile";

    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
    public User get(@AuthenticationPrincipal AuthorizedUser authUser) {
        return super.get(authUser.getId());
    }

    @DeleteMapping
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void delete(@AuthenticationPrincipal AuthorizedUser authUser) {
        super.delete(authUser.getId());
    }

    @PostMapping(value = "/register", consumes = MediaType.APPLICATION_JSON_VALUE)
    @ResponseStatus(value = HttpStatus.CREATED)
    public ResponseEntity<User> register(@Valid @RequestBody UserDto userDto) {
        User created = super.create(userDto);

        URI uriOfNewResource = UserUtil.createNewURI(REST_URL, created.getId());

        return ResponseEntity.created(uriOfNewResource).body(created);
    }

    @PutMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void update(@Validated({ErrorSequence.First.class, ErrorSequence.Second.class, ErrorSequence.Third.class})
                       @RequestBody UserDto userDto, @AuthenticationPrincipal AuthorizedUser authUser) {
        super.update(userDto, authUser.getId());
    }

    @GetMapping(value = "/text")
    public String testUTF() {
        return "Русский текст";
    }
}
